requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
cssselect2==0.8.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
dj-database-url
fonttools==4.57.0
gunicorn
idna==3.10
pillow==11.2.1
python-dotenv
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
psycopg2-binary
reportlab==4.4.0
requests==2.32.3
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
